requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
passlib[bcrypt]
python-jose
dotenv
pydantic[email]
